I made a little thing
I tried to make a pacman scroller... The idea was to have the little pacman move as the page did and eat little red dots.
The theory was to have the ball move 4x the speed of the pacman meaning they would collide and the redball would reset and start to scroll again.

I couldn't figure out why it felt and scrolled janky. The ball reset and stayed in place for a good while before continuing it's scroll.